Board Game Interface - Extremely Basic
This branch contains all of the current game board that I made. The project only contains the current and easy to understand format since there is many unknown variables that is needed for the game board. So I filled in some variable inside the textbox or slots since it needs to be understandable for the other in the game team project. Also this project will need to have or find a way to enable the textblock to be visable or invisible when the mouse is hover over the textbox or slots (to show the user the amount of pieces/gems/stone/rocks in a form of number). I'll try to mess with the property and mess with the Tooltips to see if I can make either the textblock to be visible or something similar.