Images And Gameboard
I created the game pieces and added 9 to each playable box and 15 to each players box. The images will be coded to be added and taken away according to the number associated with the current box it is in. I am just adding this now so I dont lose any of the images work with the xaml code because it was a very long and tedious job.